+++ Projects/MultiphysicsSimulation/main_tese.cpp
@@ -169,11 +169,11 @@ TPZGeoMesh *MalhaGeom()
 	gmesh->NodeVec().Resize(Qnodes);
 	TPZVec<TPZGeoNode> Node(Qnodes);
 	
-	TPZVec <int> TopolQuad(4);
-	TPZVec <int> TopolLine(2);
+	TPZVec <long> TopolQuad(4);
+	TPZVec <long> TopolLine(2);
 	
 	//indice dos nos
-	int id = 0;
+	long id = 0;
 	REAL valx;
 	REAL dx=1.;
 	for(int xi = 0; xi < Qnodes/2; xi++)
